% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEM2.R
\name{fitEM2}
\alias{fitEM2}
\title{Fitting aggregate data 2}
\usage{
fitEM2(p0, opts, obs, maxiter = 100, convcrit_nll = 0.001, nomap = TRUE)
}
\arguments{
\item{p0}{initial parameter values}

\item{opts}{options}

\item{obs}{observed data}

\item{maxiter}{maximum number of iterations}

\item{convcrit_nll}{convergence criterion for the negative log-likelihood}

\item{nomap}{single model or multiple models}
}
\value{
A fitted model
}
\description{
\code{\link[=fitEM2]{fitEM2()}} implements the Expectation-Maximization(EM) algorithm for parameter
estimation of the given aggregate data model, iterating over maximum
likelihood updates with weighted MC updates. This version of the function uses nloptr instead of optimx. TOL = 1e-10
}
\examples{
#test
}
