% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendataset.R
\name{gendataset}
\alias{gendataset}
\title{Generate a dataset}
\usage{
gendataset(opts, seed = 1, reserr = TRUE, nlmixrform = FALSE)
}
\arguments{
\item{opts}{A list of options}

\item{seed}{A seed for the random number generator}

\item{reserr}{A logical indicating whether residual error should be added}

\item{nlmixrform}{A logical indicating whether the dataset should be in nlmixr format}
}
\value{
A dataset
}
\description{
\code{\link[=gendataset]{gendataset()}} generates a dataset based on the model structure and random effects, with optional residual error.
}
\examples{
#test
}
