% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meancov.R
\name{meancov}
\alias{meancov}
\title{Generate options for optimization}
\usage{
meancov(m, wt)
}
\arguments{
\item{m}{The matrix}

\item{wt}{The weights}
}
\value{
A list with the mean and covariance
}
\description{
\code{\link[=meancov]{meancov()}} computes the mean and covariance of a matrix, optionally with weights.
}
\examples{
#test
}
