% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omegas.R
\name{omegas}
\alias{omegas}
\title{Generate a dataset}
\usage{
omegas(diag, offdiag, n_om)
}
\arguments{
\item{diag}{A value for the diagonal}

\item{offdiag}{A value for the off-diagonal}

\item{n_om}{The size of the matrix}
}
\value{
A matrix with specified diagonal and off-diagonal values.
}
\description{
\code{\link[=omegas]{omegas()}} creates a matrix with specified diagonal and off-diagonal values.
}
\examples{
#test
}
