% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEM.R
\name{plot.fitEM_result}
\alias{plot.fitEM_result}
\title{Plot diagnostics of fitEM results}
\usage{
\method{plot}{fitEM_result}(x, ...)
}
\arguments{
\item{x}{A fitted model object returned by fitEM}

\item{...}{Additional arguments (not used)}
}
\description{
Plot diagnostics of fitEM results
}
