% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIRMC.R
\name{plot.fitIRMC_result}
\alias{plot.fitIRMC_result}
\title{Plot diagnostics of fitIRMC results}
\usage{
\method{plot}{fitIRMC_result}(x, ...)
}
\arguments{
\item{x}{A fitted model object returned by fitIRMC}

\item{...}{Additional arguments (not used)}
}
\description{
Plot diagnostics of fitIRMC results
}
