% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIRMC.R
\name{print.fit_admr_result}
\alias{print.fit_admr_result}
\title{Print fitIRMC results}
\usage{
\method{print}{fit_admr_result}(x, ...)
}
\arguments{
\item{x}{A fitted model object returned by fitIRMC}

\item{...}{Additional arguments (not used)}
}
\description{
Print fitIRMC results
}
