% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEM.R
\name{print.fitEM_result}
\alias{print.fitEM_result}
\title{Print fitEM results}
\usage{
\method{print}{fitEM_result}(x, ...)
}
\arguments{
\item{x}{A fitted model object returned by fitEM}

\item{...}{Additional arguments (not used)}
}
\description{
Print fitEM results
}
