% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedEM.R
\name{timedEM}
\alias{timedEM}
\title{Fitting aggregate data}
\usage{
timedEM(init, opts, obs, maxiter = 100, convcrit_nll = 5e-04, nomap = TRUE)
}
\arguments{
\item{init}{initial parameter values}

\item{opts}{options}

\item{obs}{observed data}

\item{maxiter}{maximum number of iterations}

\item{convcrit_nll}{convergence criterion for the negative log-likelihood}

\item{nomap}{single model or multiple models}
}
\value{
A fitted model
}
\description{
\code{\link[=timedEM]{timedEM()}} implements the Expectation-Maximization(EM) algorithm for parameter
estimation of the given aggregate data model, iterating over maximum
likelihood updates with weighted MC updates. It is used to compare the performance
of the old and new implementation of aggregate data modelling.
}
\examples{
#test
}
