% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admr-package.R
\docType{package}
\name{admr-package}
\alias{admr-package}
\alias{admr}
\title{admr: Aggregate data modelling in R}
\description{
A novel method to aggregate data and model it using a non-linear estimation techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hiddevandebeek/admr}
  \item Report bugs at \url{https://github.com/hiddevandebeek/admr/issues}
}

}
\author{
\itemize{
\strong{Maintainer:} H. van de Beek \email{h.van.de.beek@lacdr.leidenuniv.nl}

Authors:

\item H. van de Beek
\item P.A.J. Välitalo}
}
\keyword{internal}
