% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEM.R
\name{fitEM}
\alias{fitEM}
\title{Fitting aggregate data}
\usage{
fitEM(p0, opts, obs, maxiter = 100, convcrit_nll = 0.001, nomap = TRUE)
}
\arguments{
\item{p0}{initial parameter values}

\item{opts}{options}

\item{obs}{observed data}

\item{maxiter}{maximum number of iterations}

\item{convcrit_nll}{convergence criterion for the negative log-likelihood}

\item{nomap}{single model or multiple models}
}
\value{
A fitted model
}
\description{
\code{\link[=fitEM]{fitEM()}} implements the Expectation-Maximization(EM) algorithm for parameter
estimation of the given aggregate data model, iterating over maximum
likelihood updates with weighted MC updates.
}
\examples{
#test
}
