% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEM.R
\name{fitEM}
\alias{fitEM}
\title{Fitting aggregate data}
\usage{
fitEM(
  opts,
  obs,
  maxiter = 100,
  convcrit_nll = 1e-05,
  single_dataframe = TRUE,
  phase_fractions = c(0.2, 0.4, 0.2, 0.2),
  max_worse_iterations = 10,
  chains = 1,
  pertubation = 0.1,
  seed = 1
)
}
\arguments{
\item{opts}{options}

\item{obs}{observed data}

\item{maxiter}{maximum number of iterations}

\item{convcrit_nll}{convergence criterion for the negative log-likelihood}

\item{single_dataframe}{single model or multiple models}

\item{phase_fractions}{vector of phase fractions}

\item{max_worse_iterations}{maximum number of consecutive worse iterations before skipping a phase}

\item{chains}{number of chains}

\item{pertubation}{pertubation factor for the initial parameter values of each chain}

\item{seed}{random seed}
}
\value{
A fitted model
}
\description{
\code{\link[=fitEM]{fitEM()}} implements the Expectation-Maximization(EM) algorithm for parameter
estimation of the given aggregate data model, iterating over maximum
likelihood updates with weighted MC updates. This version of the function uses nloptr instead of optimx. TOL = 1e-10
}
\examples{
#test

}
