% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genfitfunc.R
\name{genfitfunc}
\alias{genfitfunc}
\title{Generate a fitting function for optimization}
\usage{
genfitfunc(opts, obs)
}
\arguments{
\item{opts}{The model options}

\item{obs}{The observed data}
}
\value{
A fitting function
}
\description{
\code{\link[=genfitfunc]{genfitfunc()}} generates a fitting function for optimization, taking observed data and model options as inputs.
}
\examples{
#test

}
