% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genopts.R
\name{genopts}
\alias{genopts}
\title{Generate options for optimization}
\usage{
genopts(
  f,
  g,
  xt,
  p,
  h,
  nsim = 1,
  n = 30,
  adist = NULL,
  interact = TRUE,
  fo_appr = (nsim < 10),
  biseq = NA,
  omega_expansion = 1,
  p_thetai = NA,
  single_betas = NA
)
}
\arguments{
\item{f}{The likelihood function}

\item{g}{The model function}

\item{xt}{The data}

\item{p}{The parameter settings}

\item{h}{The error function}

\item{nsim}{The number of simulations}

\item{n}{The number of simulations}

\item{adist}{The distribution of the random effects}

\item{interact}{Whether to include interaction terms}

\item{fo_appr}{Whether to use first-order approximation}

\item{biseq}{The sequence of random effects}

\item{omega_expansion}{The expansion factor for the omega matrix}

\item{p_thetai}{The theta matrix}

\item{single_betas}{The beta matrix}
}
\value{
A fitted model
}
\description{
\code{\link[=genopts]{genopts()}} initializes and generates core options and settings for modeling
and optimization, including random effects, simulation settings, and
likelihood approximations.
}
\examples{
#test
}
