% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_to_optim.R
\name{p_to_optim}
\alias{p_to_optim}
\title{Parameters to optimizable form}
\usage{
p_to_optim(p)
}
\arguments{
\item{p}{The parameter list}
}
\value{
optimized parameter list
}
\description{
\code{\link[=p_to_optim]{p_to_optim()}} Converts a parameter list into a form that can be optimized, returning transformed values, back-transformation functions, and derivatives.
}
\examples{
#test
}
