% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedbobyqa.R
\name{timedbobyqa}
\alias{timedbobyqa}
\title{Fitting aggregate data}
\usage{
timedbobyqa(p0, opts, obs)
}
\arguments{
\item{p0}{initial parameter values}

\item{opts}{options}

\item{obs}{observed data}
}
\value{
A fitted model
}
\description{
\code{\link[=timedbobyqa]{timedbobyqa()}} implements the bobqa algorithm for parameter estimation of the
given aggregate data model, iterating over maximum likelihood updates with MC
updates. Each iteration creates new MC samples and updates the parameter values.
It is used to compare the performance of the old and new implementation
of aggregate data modelling.
}
\examples{
#test
}
